backend/requirements.txt
django
djangorestframework
djangorestframework-xml
django-environ